Responsive Updates and Slight Aesthetic improvements
This update changes both the calendar and add assignment/import buttons to have responsive versions that switch in/out when breakpoints are reached.
More specifically tailwind's lg breakpoint.
The calendar switches to a vertical mode.
The "add assignment" and "import buttons" switch location and size.

On mobile versions, this will make the application more functional and user-friendly.