Responsive changes to title and instructions button
Gave the instructions button ("?") some left padding so that the responsive mobile version doesn't press the question mark icon against the left side of the page.

Also changed the title such that the text size scales with the size of the window. Allowing for a more responsive and useable website.