New responsive changes
Due to the nature of the previous responsive changes, there were an exceeding amount of merge conflicts.

As a result, I waited for a more refreshed version and have reimplemented the changes with new altercations, this version for the calendar and buttons is bug free (so I hope, please do test)

This change should thereby satisfy our requirements for a responsive design.

I have stashed changes for the vertical priority queue, as it is my belief that the priority queue functions better horizontally over vertically. I will wait for teammates input and critique on this decision however.